requirements.txt
qdrant-client>=1.12.0
sentence-transformers>=3.0.0
torch>=2.2.0
numpy>=1.26.0
python-dotenv>=1.0.1
openai>=1.40.0
pypdf>=4.0.0
#pip install google-generativeai
